Add email adress on user entity
