+++ app/Resources/views/form/_register_form_theme.html.twig
@@ -19,6 +19,14 @@
         </tr>
         <tr>
             <td>
+                {{ form_label(form.mail) }}
+            </td>
+            <td>
+                {{ form_widget(form.mail) }}
+            </td>
+        </tr>
+        <tr>
+            <td>
                 {{ form_label(form.userName) }}
             </td>
             <td>
